want_present/requirements.txt
anyio==4.3.0
APScheduler==3.6.3
asgiref==3.8.1
attrs==23.2.0
cachetools==4.2.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==3.2.16
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
exceptiongroup==1.2.1
flake8==6.0.0
flake8-isort==6.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.10.0
pycparser==2.22
pyflakes==3.0.1
PyJWT==2.8.0
pytest==7.0.0
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
python-dotenv==1.0.1
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.11.0
tzlocal==5.2
urllib3==2.2.1
